fix: vercel build error

+++ utils/build-tracker.js
@@ -11,9 +11,16 @@ function buildTracker() {
     return;
   }
 
+  const dist = path.resolve(__dirname, '../dist');
+  if (!fs.existsSync(dist)) {
+    fs.mkdirSync(dist);
+  }
+
   const src = path.resolve(__dirname, '../node_modules/vector-tracker/lib/vector.min.js');
   const dest = path.resolve(__dirname, `../dist/${fileName}`);
-  fs.copyFileSync(src, dest);
+  if (fs.existsSync(src)) {
+    fs.copyFileSync(src, dest);
+  }
 }
 
 module.exports = buildTracker;
